+++ api/src/modules/employee/employee.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from 'src/prisma/prisma.module';
-import { EmployeeController } from './employee.controller';
-import { EmployeeService } from './employee.service';
+import { EmployeeController } from './presentation/employee.controller';
+import { EmployeeService } from './application/employee.service';
 
 @Module({
   imports: [PrismaModule],
